Fixed string representations
The string representation for the digit "3" couldn't be correctly converted back to its original form upon decryption because of two identical characters next to each other. Because of this, the representation would be changed by an extra character if the two characters happened to be in the same pair when filling in extra characters to allow for conversion of the plaintext to digraphs.

By adding an "Y" character at the start of every representation, the chance of leaving extra characters behind after decryption is reduced.